+++ src/package_managers/mod.rs
@@ -1,1 +1,2 @@
 pub(crate) mod npm;
+pub(crate) mod pnpm;